refactor: repository Directory 구조 수정
