+++ sesac7_flutter_recipe_app/lib/core/routing/router.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_recipe_app/core/routing/routes.dart';
 import 'package:flutter_recipe_app/data/data_source/recipe_data_source/recipe_data_source.dart';
 import 'package:flutter_recipe_app/data/data_source/recipe_data_source/recipe_data_source_impl.dart';
-import 'package:flutter_recipe_app/domain/repository/recipe_repository/recipe_repository.dart';
-import 'package:flutter_recipe_app/data/repository/recipe_repository/recipe_repository_impl.dart';
+import 'package:flutter_recipe_app/domain/repository/recipe_repository.dart';
+import 'package:flutter_recipe_app/data/repository/recipe_repository_impl.dart';
 import 'package:flutter_recipe_app/presentation/home/home_screen.dart';
 import 'package:flutter_recipe_app/presentation/main/main_screen.dart';
 import 'package:flutter_recipe_app/presentation/saved_recipe/saved_recipe_screen.dart';
