+++ sesac7_flutter_recipe_app/lib/presentation/search_recipe/search_recipes_view_model.dart
@@ -5,7 +5,7 @@ import 'package:flutter_recipe_app/core/enum/search_recipe_filter_category_type.
 import 'package:flutter_recipe_app/core/enum/search_recipe_filter_time_type.dart';
 import 'package:flutter_recipe_app/core/result.dart';
 import 'package:flutter_recipe_app/domain/model/recipe.dart';
-import 'package:flutter_recipe_app/domain/repository/recipe_repository/recipe_repository.dart';
+import 'package:flutter_recipe_app/domain/repository/recipe_repository.dart';
 import 'package:flutter_recipe_app/presentation/search_recipe/search_recipes_state.dart';
 
 class SearchRecipesViewModel with ChangeNotifier {
